requirements.txt
requests
beautifulsoup4
mysql-connector-python
python-dotenv
flask
gunicorn